requirements.txt
matplotlib>=3.1
numpy>=1.18
opencv>=3.4
pytorch>=1.3
torchvision>=0.4
tqdm>=4.38